Fix the "scee" and "scnb" actions to update the dihedrals in addition to the parm_data numbers so they will propagate to any OpenMM system that is generated after those commands are used.
Also add a "applayer" argument to the "energy" command. This will force ParmEd
to compute an energy using the amberprmtopfile.AmberPrmtopFile from the OpenMM
App layer instead of the OpenMMAmberParm included with ParmEd. This option is
included primarily so people can validate ParmEd's OpenMM integration with the
'official' option.  In all systems I've tested the results are the same.