Significant improvement to the ParmEd API. Now you can import actions via the command:
from ParmedTools import HMassRepartition, change, setBond, deleteBond

or

from ParmedTools import *

which imports names that match the names in the manual and that are
autocompleted in the interactive interpreter. The past behavior required users
to import all-lowercase action names from ParmedTools.ParmedActions (which is
where and how they are implemented). It was implemented this way originally to
allow the ParmEd interpreter to be case-insensitive, but could be confusing.

The old way still works fine.