Remember the atomic number after the first time it was set.
This fixes issues in which someone may change some atomic masses and those
atomic numbers will be reset to strange values if ATOMIC_NUMBER is not set in
the topology file.  This change makes it so the atomic number will be set at the
beginning of the Python program and not change (unless requested to) for the
rest of the script.