Do filename globbing in the chamber action.
Also add a -toppar flag so that 'chamber' will try to determine what the file
type of each file is based on extension (or the presence of 'top' or 'par' in
the case that the extension is '.inp'). This should make input simpler.

Add a test for the new globbing functionality.