Support the CHARMM-specific SEGID identifier in the PDB parser. It assigns it to an attribute on the Atom instance (it should probably really be on the 'residue', but this is a hack anyway and putting it on the atom allows the reader/writer to faithfully mimic each other when a single residue has atoms with multiple SEGIDs -- not sure if this should be an error instead).
Also provide a charmm flag for the PDB writer to optionally spit out the SEGID
in the 73rd through 76th columns (4 characters).  The "segid" is really what I
was calling the "system" in the PSF parser (and the 'system' is what I was
assigning to the residue chain).  Move this to the atom-based SEGID, which
should improve the PDB file saving behavior of what CHARMM users expect.