Attempt to prevent random Travis failures
I presume these failures are caused by blowing the memory limits placed on
tests. I think this is caused by rogue System objects and Context objects
sticking around in memory. First stab to get this working is to just call a
gc.collect to try and force the garbage collector to clean up those objects.

If need be, I can del them, but I'd rather not have to.