Fix the state data reporters and their tests
Awhile ago a problem with the reporter tests in which the class names changed
was masked as a "OpenMM isn't present, so don't run the tests" issue, so bugs in
the reporter classes were never caught (since the tests were never run). This
commit fixes the issue with has_numpy not being in the geometry test namespace
as well as the fact that the OpenMM reporters were not being tested and were
broken.