Fix and improve mol2 file writing.
Fixes mol2 file writing so it works with ResidueTemplateContainer objects as
well as Structure instances.  Still needs some improvement with mol_name
handling (mol2 files offer too many different names for their units... sigh)

It also puts the output in a more tabular format that is easy to grok.  This
format will overflow slightly with over 1M atoms or very long atom names (or
atom type names), but it will still maintain its whitespace delimited behavior
(it just won't look quite as tabular).  But for any files that large, grokking
by eye is unlikely to be nearly as important (and it won't look horrible, it
just won't look as good).