A number of steps toward supporting Gromacs topology files. The way that the gromacs topology works is by defining a topology for all of the individual "molecules" in the system in separate sections, then giving a "molecules" section at the end that tells how many of each kind of molecule is present.  So this necessitates the capability to combine and replicate Structure, which I implemented as the addition and multiplication operators (and their in-place variants).
I tried to do this with efficiency in mind (in particular, I was keen to keep
the method linearly scaling with the size of the structure and number of
replicates). This also necessitated the ability to duplicate parameter types, so
I added __copy__ methods to each of them.

The existing tests pass, I still need to add tests for the added functionality.