Make the "split()" function always (only) return unique molecules, as that's the only use-case that I currently have for it.
It is now *MUCH* faster than the original implementation, to the point where
it's actually useful (first approaches were ~60-120 seconds on a system that now
takes 2 seconds).

And as long as there are not a lot of unique molecule types, the cost does not
increase much with increasing system size.