Add missing test case. Also add a better default GROMACS_TOPDIR
If GROMACS_TOPDIR can't be determined from any of the candidates GMXBIN,
GMXDATA, `which gmx`, or `which mdrun`, set it to
/usr/local/gromacs/share/gromacs, since that is where it gets installed to by
default if users install it from source.

Also note that /usr/share/gromacs (among common variants) are searched before
this fallback, which catches all of the package managers that I'm aware of
(including MacPorts).