First stab at writing parameters to the Gromacs topology file.
Implemented the full proposed interface:

GromacsTopologyFile.write(topfile, parameters='inline'|parfile|topfile)

where parameters == 'inline' means write all parameters in the same line as the
valence terms. when parameters is equal to topfile, it writes all parameters in
[ parametertype ] sections (and parameters that don't match those in the
database inline).  When parameters is a different file object or name, a
separate ITP is written and its file name is included in the generated topology
(unless the ITP is a file object whose name we don't know, then it cannot be
included).