Improve behavior of *parsing* Gromacs topology files with multiterm dihedrals specified in-line.
The issue here basically is that in-line multiterm dihedrals appear as multiple
dihedral lines in the `[ dihedrals ]` section, which used to inflate the number
of dihedral entries in the `struct.dihedrals` list compared to what you get when
you have a separate `[ dihedraltypes ]` section. This change will compress all
of the terms in a multiterm dihedral expansion specified in different lines
in-line in the `[ dihedrals ]` section into a single `DihedralTypeList` with all
of its parameters.

Duplicate periodicities are tagged as illegal, since that kind of situation is
not allowed in any other programs I'm familiar with. At this point, round-trip
reading/writing "works" (as defined in the test `test_chemistry_gromacs.py` for
reading/writing). But I still need to validate parameter equality (which would,
in turn, ensure equal energies and forces).