Fix assignment of `defaults` attribute in GromacsTopologyFile
This also adds a "round-trip" test to make sure that a written Gromacs topology
file with inline parameters gives the same Structure when read in (by comparing
all atom properties and making sure that the same valence terms are present in
the same order with the same atoms and equivalent (or very nearly equivalent)
parameter types.

This *seems* to work now.

Still TODO:

- Check the option of writing parameter types in the same file as the topology,
  but not inline
- Check the option of writing a separate ITP with the parameters and
  #include-ing that file