Make sure coordinates attribute is *always* up-to-date.
If someone changed atom.xx, xy, or xz, the coordinates could become out-of-date
if any cached versions were around.  This simplifies the code, at the expense of
some computational efficiency, but it's better to be correct and slow than
potentially incorrect and fast.

And you can still be almost just as fast, you just have to store a reference to
the Structure.coordinates and work on that, rather than the coordinates
attribute directly.  So:

coords = struct.coordinates
... do stuff with coords

rather than

... do stuff with struct.coordinates