Add Structure.save function
This allows you to write the desired format in a *single* place, with any
necessary conversions being done for you automatically. It also handles
Amber-style formats as well as possible, determining in most instances whether a
structure can be cast as a ChamberParm when AmberParm is not flexible enough.
File type is determined from filename extension *or* explicit direction.

This also fixes writing PSF files from Structure instances that were *not*
parsed from PSFs (since the `groups` list did not have a `nst2` attribute.