*Always* remake the parm data when calling write_parm
Previously, this was only called if any of the parameter arrays detected
changes, such as deleting an item (like a bond, angle, dihedral, etc.) or one of
the parameter types.  However, *changing* one of the parameter types would not
trigger AmberParm to recognize a change, so the raw data would not be rebuilt
upon writing.  Since this is probably fairly common, we might as well *always*
flush updates to the raw data, even if it's not necessary.