Fix list assignment for AmoebaNonbondedExceptionType
Also update the nma.parm7 test file with a version that ParmEd generates. The
only difference that this has with the last one was that NMXRS (number of atoms
in the largest residue) and the SOLVENT_POINTERS section are both set correctly.
tinker_to_amber just sets these values to 0 since they aren't used for much, but
it's better to have them be correct, I'd say.