Implements .save for ResidueTemplateContainer
Supports saving to multi-molecule mol2/mol3 as well as Amber OFF file.

Also fixes potential issue in which ResidueTemplate.fix_charges. If the provided
target charge (i.e., the ``to`` keyword) has more decimal places than the
requested ``precision``, then the resulting charge of the first atom would have
higher than the desired precision. This rounds the target charge to the
requested precision so that the charges come out as close as possible to the
desired charge with the desired precision.