Drastically improve the speed of selecting a single atom from the forms:
struct[resid, atomid]
struct[chainid, resid, atomid]

By special-casing these selections. They need to be special-cased, since these
should be the only instances (along with struct[int]) that return an Atom. The
rest should return a Structure instance, even if it only selects a single atom.

Also fixes a bug in chain selections in cases where chains are not contiguously
listed (e.g., when solvent is chain A, but there are other chains between chain
A and the solvent). Adds tests for these.