Move write directory creation into a subclassable class in utils and convert all test classes that write files into a subclass of this new FileIOTest class.
Also add tests for the genopen function. It's becoming the standard way in which
files are opened in ParmEd, and is where new file-like support (such as webpage
opening) should be added.  This also adds a test for opening a webpage, but that
code is not yet added (test-driven-development and all that).