Fixed bug when identifying CHARMM .inp files as par or top
Older CHARMM input filenames may have the form 'top*.inp' or 'par*.inp'.
Previously, the type of a *.inp file was determined based on the presence of
'top' or 'par' in the full path, rather than the filename exclusively. This led
to problems when loding files such as '/*/parmed/*/top*.inp', as encountered in
testChamberGlobbing.