Special-case natom and hasbox keywords in load_file
Some coordinate file types (e.g., Amber ASCII trajectories) require natom and
hasbox to know whether or not that information is present in the file. Just like
the ``structure`` keyword, these keywords are not required for all coordinate
file types, so let them pass silently if they're not needed, but given.