Add a load_dataframe method
This commit does a couple things. Major thing it does is add a load_dataframe
instance to Structure so that atomic properties can be set via a DataFrame (it
also works with dicts through duck typing).

This commit *also* adds a to_dataframe method to StructureView and a
load_dataframe method to StructureView