Fixes informed by pyflakes
This fixes a couple subtle bugs from missing imports and a bunch of other minor
things tagged by pyflakes. It also adds a pyflakes check to the Travis CI script
that will fail if any issues are flagged. However, pyflakes flags a lot of
issues that are not really issues, and are IMO the best way to do something. So
there is a wrapper script around pyflakes that filters out these "OK" failures
(including any failure with 3rd-party libraries that I don't want to modify
heavily).