Define a protocol for dealing with electrostatic constant
To improve the agreement of interconverted charges, ParmEd now stores charges
internally scaled to its own definition of the electrostatic constant
(calculated from values in the CRC with the highest possible precision). Input
charges read from the different sources are scaled by the ratio of the constant
used in the parent program to the ParmEd internal constant, and are scaled
*back* upon writing.

This behavior, and the electrostatic constants used in each program supported by
ParmEd, are added to the documentation in a fairly detailed section.