Add a lot more Amber mask tests.
Also extend the mask parser a little bit. The characters @/ seemed to have been
reserved for selecting atoms of a particular *element*, rather than having a
particular name or something. This commit adds that element support using
atomic number information stored in the relevant files.

Add a section to the docs about the addition and improve test coverage of the
Amber mask parser.

Also adds a little better error catching for foolish masks so the exception is a
little more informative.