Fix OFF writing for residues with many connections
The connection list in OFF files is only 6 characters, and the 1st two elements
are reserved for the head and tail atoms. In very rare cases where there are too
many connections, this may throw an IndexError. This fixes the IndexError by not
writing the atoms that don't "fit" in the line.