Fix an obscure bug when converting to Amber topology in some cases.
This gets rid of the dumb ``signs`` attribute that I had completely forgotten I
added (one of the early vestiges left over from the original code so many years
ago). It now uses ignore_end and improper to get the signs to be added to the
Amber topology file.

Also delete a note that is no longer accurate.