Adjust output ordering for CHARMM impropers
In particular, the periodic impropers (of the kind that Amber uses)
seems to have a somewhat peculiar order requirement, but the central
atom has to appear third.

Ensure that happens, but print out every permutation of the non-central
atoms to make sure we don't miss any parameters in the CHARMM output.