Sorting keys is a bad idea!!!
The order as written in a prm file shouldn't change. I ran into this myself where I was merging together multiple stream files for my own convenience, and noticing that my simulations now complained of missing improper parameter types. You can reverse the order of atoms with no effect, but not sort an improper by the dictionary order of the atomtypes.