Complete overhaul of the PDB parsing routines.
It now essentially discards any efforts made to conform to the PDB
standard while parsing PDB files that very clearly break that standard
(e.g., by overflowing the atom and residue index fields).

No attempt is made anymore to decode differently-encoded residue or atom
numbers (all attempts were always sequential, anyway).

Nobody should be negatively impacted by this change, and it dramatically
simplifies the routine and fixes other deficiencies. That is, it is now
a more-compliant parser than it was for PDB files that obey the
standard.

Now it's time to spend a bajillion years debugging :-/