Add DL poly writers (#1146)
* Adding some spaces

* Revert "Adding some spaces"

This reverts commit 9dd25779a935bd50117fe2ca32d0658f97c4a9da.

* Initial set of modifications to be able to convert Amber input files
into DLPOLY input files. In this commit I set the structural basis
for this changes (placed new files at parmed/dlpoly which were
adapted from parmed/gromacs), and I have an initial version of
the dlpolyconfig.py file that is already giving the expected
results. I still need to address the generation of the FIELD file
and other details

* Fix issue (that was very hard to understand and correct!) so that we now
can get the velocities from the Amber inpcrd file, in case it is there.

* Intermediate progress in making the DLPOLY FIELD file

* Complete implementation of the FIELD file generation for DLPOLY.

* Final tweaks in the formatting of the CONFIG and FIELD DLPOLY files

* Cleanup unecessary files and unecessary functions from the dlpolyfield.py file

* Add tests for the new DLPOLY conversion functionality

* Update README.md

* A few more cleanups on the file dlpolyfield.py

* Place -1 in some reshape calls inside amber/netcdffiles.py

* Remove unnecessary calls, and place relative imports

* Refactor DLPoly code to eliminate duplicate code from GROMACS

Co-authored-by: Vinicius Wilian D. Cruzeiro <vwcruzeiro@ufl.edu>