Avoid possibly duplicating constraints
This is especially possible in water where a H-H bond may be added for
constraints, but that may exist alongside a H-O-H angle. If both exist,
the current code will add the same constraint twice which could cause
problems. Now the angle constraint only adds a distance constraint if
those two atoms are not *already* constrained.

This fix is inspired from an OpenMM commit for the CHARMM files.