Add explicit hydrogens when loading in SMILES
Previously, if a user decided to use `parmed.rdkit.load_smiles` to
generate a `parmed` structure, no explicit hydrogens are added to the
structure, this can be an issue if trying to generate hydrocarbon
systems.

This PR leverages the built-in `rdkit` ability to generate explicit
hydrogens for an rdkit mol using `rdkit.Chem.AddHs` by passing a boolean
flag `hydrogens=True` to use this method. Since this could break others'
workflows, the default value is `False` to comply with current
functionality.