Prototype for heterogenous mass repartitioning
This is a draft implementation of the two-tier mass repartitioning
strategy recently published by the Sugita group. The main change is that
an additional "alternate" target hydrogen mass is supplied for hydrogens
that are connected to a ring. The complete strategy proposed for AMBER
force fields is:

1. reduce the target mass from 3*1.008 to 2.5*1.008
2. use an alternate target mass of 2*1.008

Changes to the code:

- A new optional keyword `altmass` has been added to the HMassRepartition
action that supplies the alternative mass.
- A quick scheme to produce rdkit mol objects from single-residue ParmEd
molecules has been added so that we can access the IsInRing() method for
determining which atoms are in rings.

Things to improve:

The use of rdkit molecules seems to work fine in practice, but is
obviously terribly inefficient -- a new molecule is created for each and
every atom to check rather than the other way around. I can think of
multiple ways to restructure this, but I'll kick that to the development
community. The current implementation seems to work very robustly
regardless.