Prototype for heterogenous mass repartitioning (#1186)
* Prototype for heterogenous mass repartitioning

This is a draft implementation of the two-tier mass repartitioning
strategy recently published by the Sugita group. The main change is that
an additional "alternate" target hydrogen mass is supplied for hydrogens
that are connected to a ring. The complete strategy proposed for AMBER
force fields is:

1. reduce the target mass from 3*1.008 to 2.5*1.008
2. use an alternate target mass of 2*1.008

Changes to the code:

- A new optional keyword `altmass` has been added to the HMassRepartition
action that supplies the alternative mass.
- A quick scheme to produce rdkit mol objects from single-residue ParmEd
molecules has been added so that we can access the IsInRing() method for
determining which atoms are in rings.

Things to improve:

The use of rdkit molecules seems to work fine in practice, but is
obviously terribly inefficient -- a new molecule is created for each and
every atom to check rather than the other way around. I can think of
multiple ways to restructure this, but I'll kick that to the development
community. The current implementation seems to work very robustly
regardless.

* Fix up StringIO conversion of rdkit mol object

Original implementation was a bit seat-of-my-pants. Remove the need to mess with sys.stdout.

* Revisions to address comments

- various cosmetic changes (idx instead of number attribute)
- altmass is now ring_hmass
- rdkit is now optional and a warning is raised if it is needed but not
present (fallback to the standard approach)
- look-up table based on residue idx should significantly speed-up
repartitioning on large proteins, etc.

* Added a single test.  Phenol.

Would like to include a method/property in the Atom class that
determines if atom.is_in_ring.

* Update parmed/tools/actions.py

Co-authored-by: Jason Swails <jason@entos.ai>

Co-authored-by: bradakstx <brian.radak@silicontx.com>
Co-authored-by: Tom Watson <thomas.watson@silicontx.com>
Co-authored-by: Tom Watson <83650910+tjwatson-stx@users.noreply.github.com>
Co-authored-by: Jason Swails <jason@entos.ai>