Add support for more CHARMM force fields (#1157)
* Add features to CHARMM files

* Adds support for LP (lone pairs) in CHARMM PSF files
* Adds a LocalCoordinatesFrame to support the CHARMM LP types
* Starts to add Drude parsing to CHARMM psf parser

* Avoid possibly duplicating constraints

This is especially possible in water where a H-H bond may be added for
constraints, but that may exist alongside a H-O-H angle. If both exist,
the current code will add the same constraint twice which could cause
problems. Now the angle constraint only adds a distance constraint if
those two atoms are not *already* constrained.

This fix is inspired from an OpenMM commit for the CHARMM files.

* Add test files for drude CHARMM systems

* Add a test case for parsing a DRUDE PSF and fix some issues with it

* Add some protection for changeradii against index errors

* Checkpoint commit

* More progress on Drude support

* More progress towards Drude force fields in CHARMM

Properly parses the drude anisotropies with relevant tests.

* Clean up a bunch of stuff and implement CHARMM PSF writing with drude and lone pairs

* Fix bug