Improve rdkit coordinate precision
It's possible to get the full precision in coordinates directly from rdkit even though PDB is used as the intermediary format (I know changing this is on the TODO list). 

The real nut to crack here is resolving bond orders - only mol2 and sdf really live or die by them.