Flip guess strategy for atomic numbers
Invert the order of guessing so that the first two letters are tried first instead of second -- maybe there's a good reason why this wasn't done this way originally?

This lead to a bug in my code when a mol2 from AmberTools Antechamber was generated with the name/type CL1 and ParmEd kept reading it as carbon. This change fixes my case at least, but maybe it breaks something else I'm not thinking about?